requirements.txt
# Web framework
fastapi>=0.110,<1.0
starlette>=0.50,<1.0

# ASGI server
uvicorn[standard]>=0.27,<1.0

# Database ORM (SQLAlchemy 2.x)
SQLAlchemy>=2.0,<3.0
greenlet>=3.3,<4.0
psycopg>=3.1,<4.0

# Data validation
pydantic>=2.5,<3.0

# Config & env
python-dotenv>=1.0,<2.0

# Redis
redis>=7.1,<8.0
